+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Unreleased
 
 ### Added
+- `Entity::DANGLING` convenience constant
+
+# 0.7.6
+
+### Added
 - `World::take` for moving complete entities between worlds
+- `CommandBuffer::remove` and `CommandBuffer::despawn`
+
+### Fixed
+- Panics on use of cloned `BuiltEntityClone`s
 
 # 0.7.5